% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aion-package.R
\docType{package}
\name{ArchaeoCal-package}
\alias{ArchaeoCal-package}
\alias{_PACKAGE}
\alias{ArchaeoCal}
\title{ArchaeoCal: An R Interface to 'OxCal'}
\description{
What the package does (one paragraph).
}
\details{
\tabular{ll}{
\strong{Package:} \tab ArchaeoCal \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://archaeostat.github.io/archaeocal/}
  \item Report bugs at \url{https://github.com/tesselle/ArchaeoCal/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
33607 Pessac cedex\cr
France
}
\keyword{internal}
