% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.data.frame.OxCalResults}
\alias{as.data.frame.OxCalResults}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{OxCalResults}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{list}} returned by \code{\link[=oxcal_parse]{oxcal_parse()}}.}

\item{row.names}{A \code{\link{character}} vector giving the row names for the data
frame description, or \code{NULL}.}

\item{optional}{A \code{\link{logical}} scalar. If \code{FALSE} then the names of the
variables in the data frame are checked to ensure that they are
syntactically valid variable names and are not duplicated.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{data.frame}} with the following columns:
\describe{
\item{\code{name}}{}
\item{\code{operation}}{}
\item{\code{type}}{}
\item{\code{date}}{}
\item{\code{error}}{}
\item{\code{agreement}}{}
\item{\code{convergence}}{}
\item{\code{like_range}}{A \code{list} giving the confidence intervals of the likelihood.}
\item{\code{like_density}}{A \code{list} giving the likelihood.}
\item{\code{post_range}}{A \code{list} giving the confidence intervals of the posterior probability.}
\item{\code{post_density}}{A \code{list} giving the posterior probability.}
}
}
\description{
Coerce to a Data Frame
}
