% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal_calibrate.R
\name{oxcal_calibrate}
\alias{oxcal_calibrate}
\title{14C Calibration with OxCal}
\usage{
oxcal_calibrate(names, dates, errors, curve = "IntCal20")
}
\arguments{
\item{names}{A \code{\link{character}} vector specifying the names of the dates (e.g.
laboratory codes).}

\item{dates}{A \code{\link{numeric}} vector giving the BP dates to be calibrated.}

\item{errors}{A \code{\link{numeric}} vector giving the standard deviation of the dates
to be calibrated.}

\item{curve}{A \code{\link{character}} string specifying the calibration curve to be
used.}
}
\value{
A \code{\link{list}} with class \code{OxCalResults} (see \code{\link[=oxcal_parse]{oxcal_parse()}}).
}
\description{
14C Calibration with OxCal
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Calibrate 14C dates
cal <- oxcal_calibrate(
  names = c("X","Y"),
  dates = c(5000, 4500),
  errors = c(45, 35)
)

as.data.frame(cal)
}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_configure}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_install}()},
\code{\link{oxcal_parse}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
