% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal_execute.R
\name{oxcal_execute}
\alias{oxcal_execute}
\title{Execute an Oxcal Script}
\usage{
oxcal_execute(
  script,
  file,
  path = oxcal_path(),
  mcmc = NULL,
  verbose = getOption("ArchaeoCal.verbose"),
  ...
)
}
\arguments{
\item{script}{A \code{\link{character}} string of instructions for OxCal.}

\item{file}{A \code{\link{character}} string naming a file (without extension) to
write \code{script} to. Output files will be named after \code{file} and written to
the same directory.}

\item{path}{A \code{\link{character}} string giving the path to OxCal executable.}

\item{mcmc}{A \code{\link{character}} string giving the name of the output file for
the MCMC samples (without extension). It must match the \code{Name} argument of
OxCal's \href{https://intchron.org/tools/oxcalhelp/hlp_commands.html}{\code{MCMC_Sample()}}
function. Only used if \code{script} contains the \code{MCMC_Sample()} command.}

\item{verbose}{A \code{\link{logical}} scalar: should \R report extra information on
progress?}

\item{...}{Further parameters to be passed to \code{\link[=system2]{system2()}}.}
}
\value{
A \code{\link{list}} with class \code{OxCalOutput} containing the following elements:
\describe{
\item{\code{oxcal}}{A \code{\link{character}} string giving the path to the .oxcal file.}
\item{\code{js}}{A \code{\link{character}} string giving the path to the .js file.}
\item{\code{log}}{A \code{\link{character}} string giving the path to the .log file.}
\item{\code{txt}}{A \code{\link{character}} string giving the path to the .txt file.}
\item{\code{csv}}{A \code{\link{character}} string giving the path to the .csv file.}
}
}
\description{
Execute an Oxcal Script
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

as.data.frame(res)
}
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_file.html}
}
\author{
N. Frerebeau
}
