% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal_install.R
\name{oxcal_install}
\alias{oxcal_install}
\title{Download OxCal}
\usage{
oxcal_install(
  install_url = NULL,
  install_location = NULL,
  verbose = getOption("ArchaeoCal.verbose")
)
}
\arguments{
\item{install_url}{A \code{\link{character}} string specifying the url from which
OxCal could be installed.}

\item{install_location}{A \code{\link{character}} string specifying the directory to
extract OxCal files to. It will be created if necessary
(see \code{\link[utils:unzip]{utils::unzip()}}).}

\item{verbose}{A \code{\link{logical}} scalar: should status updates be displayed?}
}
\value{
Invisibly returns the path to the OxCal directory.
}
\description{
Download OxCal
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

as.data.frame(res)
}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_configure}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_parse}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
