% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal_parse.R
\name{oxcal_parse}
\alias{oxcal_parse}
\alias{oxcal_parse.OxCalOutput}
\title{Read and Parse OxCal Output}
\usage{
oxcal_parse(path)

\method{oxcal_parse}{OxCalOutput}(path)
}
\arguments{
\item{path}{A \code{\link{character}} string naming a JavaScript file which the data
are to be read from (or a \code{list} returned by \code{\link[=oxcal_execute]{oxcal_execute()}}).}
}
\value{
A \code{\link{list}} with class \code{OxCalResults} containing the following elements:
\describe{
\item{\code{ocd}}{A \code{\link{list}} of OxCal data which holds the ranges, probability
distributions, etc. for each parameter.}
\item{\code{model}}{A \code{\link{list}} of information about the model.}
\item{\code{calib}}{A \code{\link{list}} of information about the calibration curve.}
}
}
\description{
Read and Parse OxCal Output
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

as.data.frame(res)
}
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_file.html}
}
\author{
N. Frerebeau
}
