% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal_configure.R
\name{oxcal_path}
\alias{oxcal_path}
\title{Get OxCal Executable Path}
\usage{
oxcal_path(...)
}
\arguments{
\item{...}{Currently not used.}
}
\value{
A \code{\link{character}} string giving the path to OxCal executable.
}
\description{
Get OxCal Executable Path
}
\details{
Defaults to the R user directory.
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

as.data.frame(res)
}
}
\seealso{
\code{\link[tools:userdir]{tools::R_user_dir()}}
}
\author{
N. Frerebeau
}
