% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OxCalFiles-class}
\alias{OxCalFiles-class}
\alias{.OxCalFiles}
\title{OxCal Output Files}
\arguments{
\item{oxcal}{A \code{\link{character}} string giving the path to the .oxcal file.}

\item{js}{A \code{\link{character}} string giving the path to the .js file.}

\item{log}{A \code{\link{character}} string giving the path to the .log file.}

\item{txt}{A \code{\link{character}} string giving the path to the .txt file.}

\item{csv}{A \code{\link{character}} string giving the path to the .csv file.}
}
\description{
An S4 class to represent OxCal output files.
}
\seealso{
Other classes: 
\code{\link{OxCalOutput-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
