% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OxCalOutput-class}
\alias{OxCalOutput-class}
\alias{.OxCalOutput}
\title{OxCal Output}
\arguments{
\item{ocd}{A \code{\link{list}} of OxCal data which holds the ranges, probability
distributions, etc. for each parameter.}

\item{model}{A \code{\link{list}} of information about the model.}

\item{calib}{A \code{\link{list}} of information about the calibration curve.}
}
\description{
An S4 class to represent OxCal output.
}
\section{Coerce}{

In the code snippets below, \code{x} is an \code{OxCalOutput} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes: 
\code{\link{OxCalFiles-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
