% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoCal-package.R
\docType{package}
\name{ArchaeoCal-package}
\alias{ArchaeoCal-package}
\alias{ArchaeoCal}
\title{ArchaeoCal: An R Interface to 'OxCal'}
\description{
What the package does (one paragraph).
}
\details{
\tabular{ll}{
\strong{Version:} \tab 0.0.0 \cr
\strong{License:} \tab GPL-3 \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\section{Package options}{

\pkg{ArchaeoCal} uses the following \code{\link[=options]{options()}} to configure behaviour:
\itemize{
\item \code{ArchaeoCal.verbose}: a \code{\link{logical}} scalar. Should \R report extra
information on progress? Defaults to \code{\link[=interactive]{interactive()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://archaeostat.github.io/archaeocal/}
  \item Report bugs at \url{https://github.com/ArchaeoStat/ArchaeoCal/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})

}
\keyword{internal}
