% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal_configure.R
\name{oxcal_configure}
\alias{oxcal_configure}
\title{Configure OxCal}
\usage{
oxcal_configure(
  path = oxcal_path(),
  os = NULL,
  install = TRUE,
  ask = interactive(),
  force = FALSE,
  verbose = getOption("ArchaeoCal.verbose"),
  ...
)
}
\arguments{
\item{path}{A \code{\link{character}} string specifying the path to the OxCal
directory.}

\item{os}{A \code{\link{character}} string specifying the operating system of the
workstation. It must be one of "\code{Linux}", "\code{Windows}" or "\code{Darwin}".
If \code{NULL} (the default), the operating system will be determined
automatically (see \code{\link[=Sys.info]{Sys.info()}}).}

\item{install}{A \code{\link{logical}} scalar: if OxCal binary is not found,
should it be downloaded and installed?}

\item{ask}{A \code{\link{logical}} scalar: if OxCal is not installed, should the user
be asked before downloading it?}

\item{force}{A \code{\link{logical}} scalar: if OxCal is already installed, should
reinstallation be forced?}

\item{verbose}{A \code{\link{logical}} scalar: should \R report extra information on
progress?}

\item{...}{Further arguments to be passed to \code{\link[=oxcal_install]{oxcal_install()}}.}
}
\value{
\code{oxcal_configure()} is called for its side-effect.
Invisibly returns the path to the OxCal executable.
}
\description{
Configure OxCal
}
\details{
Downloads the latest version of Oxcal (if needed) and sets the executable
path correctly.
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

as.data.frame(res)
}
}
\author{
N. Frerebeau
}
