% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal_execute.R
\name{oxcal_execute}
\alias{oxcal_execute}
\title{Execute an Oxcal Script}
\usage{
oxcal_execute(script, file = NULL, verbose = getOption("aion.verbose"), ...)
}
\arguments{
\item{script}{A \code{\link{character}} string of instructions for OxCal.}

\item{file}{A \code{\link{character}} string naming a file (without extension) to
write \code{script} to. Output files will be named after \code{file} and written to
the same directory.}

\item{verbose}{A \code{\link{logical}} scalar: should status updates be displayed?}

\item{...}{Further parameters to be passed to \code{\link[=system2]{system2()}}.}
}
\value{
An \code{\linkS4class{OxCalFiles}} object.
}
\description{
Execute an Oxcal Script
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

plot(res)
plot(res, likelihood = TRUE, posterior = FALSE)
plot(res, likelihood = FALSE, posterior = TRUE)
}
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_file.html}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_configure}()},
\code{\link{oxcal_install}()},
\code{\link{oxcal_parse}()},
\code{\link{plot}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
