% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal_install.R
\name{oxcal_install}
\alias{oxcal_install}
\title{Download and Install OxCal}
\usage{
oxcal_install(
  target,
  source = "https://c14.arch.ox.ac.uk/OxCalDistribution.zip",
  verbose = getOption("ArchaeoCal.verbose"),
  ...
)
}
\arguments{
\item{target}{A \code{\link{character}} string specifying the directory to
extract OxCal files to. It will be created if necessary
(see \code{\link[utils:unzip]{utils::unzip()}}).}

\item{source}{A \code{\link{character}} string specifying the url from which
OxCal could be installed.}

\item{verbose}{A \code{\link{logical}} scalar: should \R report extra information on
progress?}

\item{...}{Currently not used.}
}
\value{
\code{oxcal_install()} is called for its side-effect.
Invisibly returns the path to the OxCal directory.
}
\description{
Download and Install OxCal
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

as.data.frame(res)
}
}
\author{
N. Frerebeau
}
