% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/oxcal_parse.R
\docType{methods}
\name{oxcal_parse}
\alias{oxcal_parse}
\alias{oxcal_parse-method}
\alias{oxcal_parse,OxCalFiles-method}
\alias{oxcal_parse,character-method}
\title{Read and Parse OxCal Output}
\usage{
oxcal_parse(object)

\S4method{oxcal_parse}{OxCalFiles}(object)

\S4method{oxcal_parse}{character}(object)
}
\arguments{
\item{object}{A \code{\link{character}} string naming a JavaScript file which the data
are to be read from (or an \code{\linkS4class{OxCalFiles}} object returned by
\code{\link[=oxcal_execute]{oxcal_execute()}}).}
}
\value{
An \code{\linkS4class{OxCalOutput}} object.
}
\description{
Read and Parse OxCal Output
}
\examples{
\dontrun{
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

plot(res)
plot(res, likelihood = TRUE, posterior = FALSE)
plot(res, likelihood = FALSE, posterior = TRUE)
}
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_file.html}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_configure}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_install}()},
\code{\link{plot}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
